requirements.txt
django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2==2.8.6
gunicorn==20.1.0
django-heroku==0.3.1
django_cloudinary_storage==0.3.0
pillow==8.3.2
django-cors-headers==3.9.0
PyJWT==2.1.0